+++ examples/oneapi/rate_limit_oneapi.py
@@ -1,23 +1,21 @@
 """ 
-Basic example of scraping pipeline using SmartScraper
+Basic example of scraping pipeline using SmartScraper with a custom rate limit
 """
-
 from scrapegraphai.graphs import SmartScraperGraph
 from scrapegraphai.utils import prettify_exec_info
-
 # ************************************************
 # Define the configuration for the graph
-# *********************************************
+# ************************************************
+
 
 graph_config = {
     "llm": {
         "api_key": "***************************",
         "model": "oneapi/qwen-turbo",
         "base_url": "http://127.0.0.1:3000/v1",  # 设置 OneAPI URL
-    },
-    "embeddings": {
-        "model": "ollama/nomic-embed-text",
-        "base_url": "http://127.0.0.1:11434",  # 设置 Ollama URL
+        "rate_limit": {
+            "requests_per_second": 1
+        }
     }
 }
 
@@ -26,15 +24,18 @@ graph_config = {
 # ************************************************
 
 smart_scraper_graph = SmartScraperGraph(
-    prompt="该网站为XXXXX,请提取出标题、发布时间、发布来源以及内容摘要,并以中文回答。",
-    # 也可以使用已下载的 HTML 代码的字符串
-    source="http://XXXX",
+    prompt="List me all the titles",
+    # also accepts a string with the already downloaded HTML code
+    source="https://www.wired.com/",
     config=graph_config
 )
+
+result = smart_scraper_graph.run()
+print(result)
 
 # ************************************************
 # Get graph execution info
 # ************************************************
-result = smart_scraper_graph.run()
-print(result)
-print(prettify_exec_info(result))
+
+graph_exec_info = smart_scraper_graph.get_execution_info()
+print(prettify_exec_info(graph_exec_info))